Kind of scuffed movement of enemy along traced path

+++ Assets/Scripts/CellStats.cs
@@ -3,8 +3,8 @@ using UnityEngine;
 public class CellStats : MonoBehaviour
 {
     public CellStats parent;
-    public int F;
-    public int G;
-    public int H;
+    public float F;
+    public float G;
+    public float H;
 
 }
